requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
argon2-cffi>=23.1
itsdangerous>=2.1
python-multipart>=0.0.9
